+++ nix/overlays/lib.nix
@@ -3,7 +3,7 @@ self: super:
 {
   lib = (super.lib or {}) // {
     maintainers = super.lib.maintainers // {
-      dhess = "Drew Hess <dhess-src@quixoftic.com>";
+      dhess = "Drew Hess <src@drewhess.com>";
     };
   };
 }