+++ src/System/GPIO.hs
@@ -1,9 +1,9 @@
 {-|
 Module      : System.GPIO
 Description : Top-level re-exports for writing GPIO programs
-Copyright   : (c) 2018, Quixoftic, LLC
+Copyright   : (c) 2018, Drew Hess
 License     : BSD3
-Maintainer  : Drew Hess <dhess-src@quixoftic.com>
+Maintainer  : Drew Hess <src@drewhess.com>
 Stability   : experimental
 Portability : non-portable