+++ changelog.md
@@ -1,5 +1,6 @@
 Changes:
 
+  - Support for Stackage LTS 12.
   - Support for GHC 8.4.3.
 
 ## 0.9.0.6 (2018-04-03)