+++ changelog.md
@@ -2,6 +2,8 @@ Changes:
 
   - Support for Stackage LTS 12.
   - Support for GHC 8.4.3.
+  - Copyright for the project has been re-assigned to Drew Hess, its
+    original author.
 
 ## 0.9.0.6 (2018-04-03)